refactor(web): extract shared DraftSaveDiffModal component
The decap PrefixSaveDiffModal and route FIBSaveDiffModal were identical
modulo row type and the module-local diff/validation helpers. Hoist the
shared logic into a generic DraftSaveDiffModal under components/ that
parameterizes the YAML serializer and the invalid-row counter, and reduce
both module wrappers to declarative pass-throughs.
